WI-57015 Inspect code with enabled PHPStan inspection spawns many PHP processes, seems never to finish
The logic of execution is the following. If FULL_PROJECT is enabled, always analyze the whole project (pass project dir to PHPStan). If not and config is provided, don't pass any path (paths should be provided in config). If not and config is not provided, pass all source roots.

(cherry picked from commit 5189681e85260505277ed9d73ed34bd60cc4a8be)

IJ-CR-4759

GitOrigin-RevId: 4a091f1d1ab7e0406ac5bf2efe83063b2458ba27